Add Anders Kristiansen from Nethermind
 ### Name
Anders Kristiansen
 
**GitHub**: https://github.com/ak88
 
 ### Team
 [Nethermind](https://github.com/NethermindEth/nethermind)
 
 ### Link to some work
 https://github.com/NethermindEth/nethermind/pulls?q=author%3Aak88
 
 ### Eligibility
Anders joined Nethermind in November 2023 as an intern. He worked on ERA1 exports/imports and contributed to a tool used for testing blobs, which helped uncover bugs in different clients. He has worked on various bug fixes that have benefited Nethermind. Anders was also the main developer for EIPs 3074 and 7702.
 
 ### Start Date
Anders joined the Nethermind team in September 2023 as an intern and became a regular member in December 2023
 
 ### Proposed Weight
 Partial. Anders will be spending part of his time working on projects not related to Ethereum L1.