Add zkEVM working group
**Proposal: Add zkEVM working group**

This proposal outlines the rationale of creating a **zkEVM working group** within the Protocol Guild. This working group will maintain a focus on foundational protocol-level work that benefits the broader ecosystem, rather than specific product or client development.

This PR acknowledges that Ethereum's ["snarkification" roadmap](https://blog.ethereum.org/2025/07/31/lean-ethereum) is still in its early stages and will likely evolve over several years. However, the launch of a zkEVM Attester Client is anticipated in the short-to-medium term:

* [Shipping an L1 zkEVM \#1: Realtime Proving](https://blog.ethereum.org/2025/07/10/realtime-proving)  
* [Protocol Update 001 – Scale L1](https://blog.ethereum.org/2025/08/05/protocol-update-001)  
* [Making Sense of a ZK Staking Node](https://paragraph.com/@ethstaker/making-sense-of-a-zk-staking-node)

Therefore, it is appropriate to include this initiative in Protocol Guild's "Wayfinding" category, which is defined as "the exploratory process to surface, describe and validate potential protocol changes". A similar longterm exploration would be post-quantum crypto, or the verkle/stateless efforts.

This PR also adds Kev to this new category, who have already been engaging in this work.

**Initial Eligible Scope**

* Includes [prototyping work](https://github.com/sigp/lighthouse/pull/7755) for a zkEVM attester client. The client will follow a [draft specification](http://github.com/ethereum/consensus-specs/pull/4591) that integrates both consensus and execution verification into a single binary.  
* May include guest program benchmarking, guest program security, zkEVM coordination, zkEVM proof verification, zkEVM specifications, and guest program compilation.

**Exclusions**

["Lean Consensus" client teams](https://leanroadmap.org) are not covered by this proposal.